+++ src/main/kotlin/io/github/jhspetersson/turtlecommander/vfs/VirtualFileSystem.kt
@@ -602,6 +602,7 @@ object VirtualFileSystemRegistry {
         register(RarFileSystemProvider())
         register(CabFileSystemProvider())
         register(ArFileSystemProvider())
+        register(ArjFileSystemProvider())
         register(XzFileSystemProvider())
         register(ZstFileSystemProvider())
         register(CrxFileSystemProvider())